pyproject.toml
[tool.poetry]
name = "earthquake-service"
version = "0.1.0"
description = ""
authors = ["Artur Gomes <contato@arturgomes.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0.4"
chardet = "4.0.0"
datadispatch = "1.0.0"
dj-database-url = "0.5.0"
django-cors-headers = "3.11.0"
django-split-settings = "1.1.0"
djangorestframework = "3.13.1"
libsass = "0.20.0"
lockfile = "0.12.2"
psycopg2-binary = "2.9.3"
PyJWT = "2.3.0"
python-dateutil = "2.8.2"
python-decouple = "3.6"
redis = "4.2.2"
requests = "2.27.1"
cryptography = "^36.0.2"
django-silk = "^4.3.0"
celery = "^5.2.6"
django-ckeditor = "^6.3.2"
Pillow = "^9.1.0"
whitenoise = "^6.0.0"

[tool.poetry.dev-dependencies]
Faker = "8.10.3"
Jinja2 = "3.0.1"
jsonschema = "3.2.0"
mixer = "7.1.2"
pylint-django = "^2.5.3"
coverage = "^6.3.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
